% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Functions.R
\name{Error_Checker}
\alias{Error_Checker}
\title{Error_Checker}
\usage{
Error_Checker(x)
}
\arguments{
\item{x}{value from try()}
}
\description{
This function allows to check for error messages from try().
}
