% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Functions.R
\name{FUN_PoI_BIC}
\alias{FUN_PoI_BIC}
\title{FUN_PoI_BIC}
\usage{
FUN_PoI_BIC(Y = Y, X.mat = X.mat, a = 0, b = 1, CoVars = NULL,
  S.max = 5, show.pb = FALSE, k.seq = NULL, IC = c("bic",
  "aic")[1], standardize.X = FALSE, center.X = FALSE,
  family = "binomial")
}
\arguments{
\item{Y}{dependent variable (n-vector)}

\item{X.mat}{pxn matrix of discretized functional predictors}

\item{a}{lower interval border (default a=0)}

\item{b}{upper interval border (default b=1)}

\item{CoVars}{additinal covariables (default CoVars=NULL)}

\item{S.max}{maximum number of points of impact (default S.mat=5)}

\item{show.pb}{show a progress bar (default show.pb = FALSE)}

\item{k.seq}{define k.seq (i.e., the delta values - in terms of grid-points - to be considered). The default (k.seq=NULL) means: k.seq = ceiling(p * seq(0.02, 0.40, by = 0.02))}

\item{IC}{information criterion ("bic" (default), or "aic")}

\item{standardize.X}{standardize (to unit variance) the functional predictors (default standardize.X  = FALSE)}

\item{center.X}{center the functional predictors (default center.X = FALSE)}

\item{family}{any family of ?stats::family (default family = "binomial")}
}
\description{
This function is an implementation of the fully data driven estimator POI.
}
