% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Functions.R
\name{FUN_PoI_TRH}
\alias{FUN_PoI_TRH}
\title{FUN_PoI_THR}
\usage{
FUN_PoI_TRH(Y, X.mat, a = 0, b = 1, scale = 1.5,
  force.1st.PoI = FALSE)
}
\arguments{
\item{Y}{dependent variable (n-vector)}

\item{X.mat}{pxn matrix of discretized functional predictors}

\item{a}{lower interval border (default a=0)}

\item{b}{upper interval border (default b=1)}

\item{scale}{scaling parameter c_delta (default scale=1.5)}

\item{force.1st.PoI}{force selection of the first impact point (default force.1st.PoI = FALSE)}
}
\description{
This function is an implementation of the threshold estimator THR.
}
