% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Functions.R
\name{OU.SIM}
\alias{OU.SIM}
\title{OU.SIM}
\usage{
OU.SIM(a, b, p, N, theta = 5, sigmaou = 3.5)
}
\arguments{
\item{a}{lower interval boundary}

\item{b}{upper interval boundary}

\item{p}{length of the path}

\item{N}{number of paths}

\item{theta}{parameter theta}

\item{sigmaou}{parameter sigma}
}
\description{
This function allows to simulate Ornstein–Uhlenbeck processes.
}
