% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Functions.R
\name{TauIndHat_to_VideoSec}
\alias{TauIndHat_to_VideoSec}
\title{TauIndHat_to_VideoSec}
\usage{
TauIndHat_to_VideoSec(TauIndHat)
}
\arguments{
\item{TauIndHat}{Selected PoI indexes from FUN_PoI_BIC() or FUN_PoI_TRH()}
}
\description{
This function allows to compute the video time points for each PoI grid-point.
}
